refactor: extract RoutinesDashboardModal component
